fix: inline scope-cache
does not work correctly with esm imports since in the resulting build
scope-cache does not have explicit "js" extension

This is hotfix, proper fix requires changes in build config
